revision
I wrote in a way that came naturally, in a way where I was pretending to be having a conversation.

I made sure not to overwrite. Doing so in this simple document could have easily caused confusion.

No opinion was injected, and the document was kept as factual as possible.